delete gtfs-rt-archive application code; keep 2 functions for agencies_verify
