--- services/gtfs-rt-archive/requirements.txt
@@ -1,12 +1,0 @@
-cachetools==4.2.2
-certifi==2021.5.30
-charset-normalizer==2.0.3
-google-auth==1.33.0
-idna==3.2
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-PyYAML==5.4.1
-requests==2.26.0
-rsa==4.7.2
-six==1.16.0
-urllib3==1.26.6
