--- services/gtfs-rt-archive/src/gtfs_rt_archive/__main__.py
@@ -1,3 +1,0 @@
-from . import main
-
-main()
